:sparkles: Add all validation functions to web3-validator
