+++ packages/web3-validator/package.json
@@ -26,7 +26,8 @@
 	},
 	"dependencies": {
 		"ajv": "^8.8.2",
-		"web3-utils": "4.0.0-alpha.0"
+		"web3-utils": "4.0.0-alpha.0",
+		"ethereum-cryptography": "^0.2.1"
 	},
 	"devDependencies": {
 		"@types/jest": "^27.0.3",
